Workaround for connect baremetal network to neutron_external_interface
Ironic conductor and TFTP server listens on IP address on interface attached
to baremetal network.
In case when baremetal network is on same interface as
neutron_external_interface, that interface should not be added to OVS bridge
deirectly.
Instead it should be added to linux bridge (with configured IP on bridge)
and that bridge interconnects with OVS bridge (br-ex) via veth pair.

Change-Id: I82dcb80e5dfffdaab44381a9aa5f382cf9a3c4b0